# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Bookstore API",
        "description": "這是書店的 API 文件",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/menu/GetMenu": {
            "get": {
                "tags": [
                    "Menu"
                ],
                "summary": "取得所有菜單項目",
                "operationId": "705497f30f71e5155f123df76c9ad1ac",
                "responses": {
                    "200": {
                        "description": "成功取得菜單",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "首頁"
                                            },
                                            "url": {
                                                "type": "string",
                                                "example": "/Home"
                                            },
                                            "icon": {
                                                "type": "string",
                                                "example": "pi-home"
                                            },
                                            "order": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "visible_for": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string"
                                                },
                                                "example": [
                                                    "user",
                                                    "admin"
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "驗證錯誤"
                    }
                }
            }
        },
        "/api/menu/CreateMenu": {
            "post": {
                "tags": [
                    "Menu"
                ],
                "summary": "新增一筆 Menu",
                "operationId": "a7f06c71a9fccfab98ea2a14513c5a00",
                "requestBody": {
                    "description": "新增菜單所需資料",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "url"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "首頁"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "/Home"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "pi-home"
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "visible_for": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "user",
                                            "admin"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "新增成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "menu_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "首頁"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "/Home"
                                        },
                                        "icon": {
                                            "type": "string",
                                            "example": "pi-home"
                                        },
                                        "order": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "visible_for": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "user",
                                                "admin"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "驗證錯誤"
                    }
                }
            }
        },
        "/api/menu/UpdateMenu/{id}": {
            "put": {
                "tags": [
                    "Menu"
                ],
                "summary": "更新特定 Menu",
                "operationId": "a1a8b7630f43bc8ee04922bfc5736119",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu 的 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "更新資料",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "回首頁"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "/Home"
                                    },
                                    "icon": {
                                        "type": "string",
                                        "example": "pi-home"
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "visible_for": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "user"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "menu_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "回首頁"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "/Home"
                                        },
                                        "icon": {
                                            "type": "string",
                                            "example": "pi-home"
                                        },
                                        "order": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "visible_for": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "user"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "驗證錯誤"
                    }
                }
            }
        },
        "/api/menu/DeleteMenu/{id}": {
            "delete": {
                "tags": [
                    "Menu"
                ],
                "summary": "刪除特定 Menu",
                "operationId": "18fa9252576d7b6369583cd9fe65dfec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Menu 的 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刪除成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "刪除成功"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "找不到指定 Menu"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}